convert js to ts
