--- frontend/disco-dj/src/App.jsx
@@ -1,21 +1,0 @@
-import { useState } from 'react'
-import disco_ball from './assets/disco_ball.png'
-import './App.css'
-
-function App() {
-  const [count, setCount] = useState(0)
-
-  return (
-    <>
-    <body className='homepage'>
-      <div className='introduction'>
-        <img src={disco_ball} className="logo react" alt="React logo" />
-        <h1>Welcome to Disco DJ</h1>
-        </div>
-    </body>    
-    </>
-
-  )
-}
-
-export default App
