+++ frontend/src/App.tsx
@@ -1,0 +1,18 @@
+// import disco_ball from './assets/disco_ball.png'
+import './App.css'
+
+function App() {
+
+  return (
+    <>
+    <body className='homepage'>
+      <div className='introduction'>
+        {/* <img src={disco_ball} className="logo react" alt="React logo" /> */}
+        <h1>Welcome to Disco DJ</h1>
+        </div>
+    </body>  
+    </>
+  )
+}
+
+export default App
